# Speed up `astropy.stats.histogram_intervals()`